CLEANUP: Remove unused functions
Basically, these will suck to support. A VFS instance should be
constructed once and considered to be immutable thereafter.

In cases such as the builtin data directories provided by openmw, these
are either prepended in the case of `vfs/data-mw` or appended in the
case of `data-local`, so this won't invalidate the accuracy of this
replication.